Make the full-service playbook more robust
Sometimes, the playbook would fail because the SSH port on the deployed
VMs would be available, but the SSH necessary to log in was not set up
yet.

The playbooks now pause until they're able to SSH into the deployed
servers.

And the Heat stack is now deployed only when it doesn't exist already.
This makes the playbook idempotent.